refactor with solid principles v3
