+++ client/src/services/socketService.js
@@ -1,0 +1,13 @@
+import { io } from 'socket.io-client'
+
+export const socketService = (data) => {
+    io('http://localhost:5000', {
+        autoConnect: false,
+        query: {
+            instaStreamURL: data.instaStreamURL,
+            instaStreamKey: data.instaStreamKey,
+            youtubeStreamURL: data.youtubeStreamURL,
+            youtubeStreamKey: data.youtubeStreamKey,
+        },
+    });
+};
